Move common.views to tests.views
