+++ starlette_web/core/routes.py
@@ -1,7 +1,8 @@
 from starlette.routing import Mount, Route
 
-from starlette_web.common.views import HealthCheckAPIView, SentryCheckAPIView
 from starlette_web.contrib.auth.routes import routes as auth_routes
+from starlette_web.tests.views import HealthCheckAPIView, SentryCheckAPIView
+
 
 routes = [
     Mount("/api", routes=auth_routes),
